backend/pyproject.toml
[project]
name = "smart_financial_solutions"
version = "0.1.0"
description = "Solution for data analysis using multi-agents with LangChain."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.116.2",
    "fastembed>=0.7.3",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain[google-genai,groq]>=0.3.27",
    "langsmith>=0.4.28",
    "markdown-pdf>=1.10",
    "mistune>=3.1.4",
    "numexpr>=2.10.2",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "pydantic-settings>=2.11.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "scikit-learn>=1.7.2",
    "tabulate>=0.9.0",
]

[tool.ruff.format]
quote-style = "single"

[dependency-groups]
dev = [
    "ipython>=8.37.0",
]
